requirements.txt
django
django-htmx
pandas
pypdf
ebooklib
bs4
django-bootstrap5
djangorestframework